Default to software emulation if target architecture != host_architecture
The qemu-kvm binary only supports the host architecture. We must
therefore use the qemu-system-<architecture> binaries when the target
architecture is not the same as the host architecture.

Although it may be possible that the host CPU may support acceleration
for a foreign architecture, I do not currently know of any chips that
support this functionality. For this reason, we default to software
emulation.