Always undefine a VM before defining it
It seems that the virt module does not take into account changes to
the templated vm.xml.j2 file. As such, changes to variables used
within this template are not propagated to libvirt i.e the output of
virsh dumpxml remains the same before and after the application of
define.