Fix name of libvirt_vm_image_cache_path variable
This was renamed to libvirt_image_cache_path in the defaults and README,
but the name referenced in the task files was still
libvirt_vm_image_cache_path. This change reverts to the original name,
libvirt_vm_image_cache_path.

Temporarily support the use of libvirt_image_cache_path until the next
major release.