Don't assume English locale
We currently check the output of 'virsh vol-info' for an expected error
message when checking for the presence of a volume. In non-English
locales, this will not work.

This change fixes the issue by assuming a failure of the command means
that the volume does not exist. If there is another cause, this should
be picked up when the volume is created.

Fixes: #51